+++ .gitignore
@@ -1,1 +1,3 @@
 .aws
+.cdsapirc
+*__pycache__*
