--- computation_user_data.txt
@@ -1,29 +1,0 @@
-Content-Type: multipart/mixed; boundary="//"
-MIME-Version: 1.0
-
---//
-Content-Type: text/cloud-config; charset="us-ascii"
-MIME-Version: 1.0
-Content-Transfer-Encoding: 7bit
-Content-Disposition: attachment; filename="cloud-config.txt"
-
-#cloud-config
-cloud_final_modules:
-- [scripts-user, always]
-
---//
-Content-Type: text/x-shellscript; charset="us-ascii"
-MIME-Version: 1.0
-Content-Transfer-Encoding: 7bit
-Content-Disposition: attachment; filename="userdata.txt"
-
-#!/bin/bash
-HOME=/home/ubuntu
-cd $HOME/retrospective-update/computation_scripts
-eval $($HOME/miniconda3/bin/conda shell.bash hook)
-source $HOME/miniconda3/bin/activate add_week
-source ../.profile
-python master.py
-sudo shutdown -h now
-
---//--
