--- downloader_scripts/environment.yaml
@@ -1,23 +1,0 @@
-name: download
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - awscli
-  - python>=3.8
-  - boto3
-  - cdsapi
-  - cftime
-  - dask
-  - distributed
-  - flox
-  - h5netcdf
-  - h5py
-  - netcdf4
-  - numpy
-  - pandas
-  - psutil
-  - s3fs
-  - xarray
-  - zarr
-  - natsort
